Fix some warning of unhandled cases in switch, unused variable or format.
git-svn-id: svn+ssh://svn.abisource.com/svnroot/abiword/trunk@31452 bcba8976-2d24-0410-9c9c-aab3bd5fdfd6

+++ src/wp/ap/xp/ap_EditMethods.cpp
@@ -11095,7 +11095,7 @@ Defun1(rdfQueryXMLIDs)
 			{
 				std::set< std::string > xmlids;
 				rdf->addRelevantIDsForPosition( xmlids, point );
-				UT_DEBUGMSG(("xmlids.sz:%d\n", xmlids.size() ));
+				UT_DEBUGMSG(("xmlids.sz:%lu\n", (long unsigned)xmlids.size() ));
 
 				sparql = PD_DocumentRDF::getSPARQL_LimitedToXMLIDList( xmlids );
 			}
